+++ node/src/components/contract_runtime/error.rs
@@ -1,4 +1,5 @@
 //! Errors that the contract runtime component may raise.
+use derive_more::From;
 use std::collections::BTreeMap;
 
 use serde::Serialize;
@@ -12,7 +13,7 @@ use casper_types::{bytesrepr, CLValueError, PublicKey, U512};
 
 use crate::{
     components::contract_runtime::ExecutionPreState,
-    types::{ExecutableBlock, InternalEraReport},
+    types::{ChunkingError, ExecutableBlock, InternalEraReport},
 };
 
 /// An error returned from mis-configuring the contract runtime component.
@@ -24,6 +25,20 @@ pub(crate) enum ConfigError {
     /// Error initializing metrics.
     #[error("failed to initialize metrics for contract runtime: {0}")]
     Prometheus(#[from] prometheus::Error),
+}
+
+/// An enum that represents all possible error conditions of a `contract_runtime` component.
+#[derive(Debug, Error, From)]
+pub(crate) enum ContractRuntimeError {
+    /// The provided serialized id cannot be deserialized properly.
+    #[error("error deserializing id: {0}")]
+    InvalidSerializedId(#[source] bincode::Error),
+    // It was not possible to get trie with the specified id
+    #[error("error retrieving trie by id: {0}")]
+    FailedToRetrieveTrieById(#[source] GlobalStateError),
+    /// Chunking error.
+    #[error("failed to chunk the data {0}")]
+    ChunkingError(#[source] ChunkingError),
 }
 
 /// An error during block execution.
