+++ node/src/components/contract_runtime/tests.rs
@@ -7,11 +7,12 @@ use serde::Serialize;
 use tempfile::TempDir;
 
 use casper_types::{
-    bytesrepr::Bytes, runtime_args, Chainspec, ChainspecRawBytes, Deploy, EraId,
-    ExecutableDeployItem, PublicKey, SecretKey, TimeDiff, U512,
+    bytesrepr::Bytes, runtime_args, BlockHash, Chainspec, ChainspecRawBytes, Deploy, Digest, EraId,
+    ExecutableDeployItem, PublicKey, SecretKey, TimeDiff, Timestamp, U512,
 };
 
 use super::*;
+use crate::types::{ExecutableBlock, MetaBlockState};
 use crate::{
     components::{
         network::Identity as NetworkIdentity,
@@ -276,7 +277,7 @@ async fn should_not_set_shared_pre_state_to_lower_block_height() {
             .execution_pre_state
             .lock()
             .unwrap()
-            .next_block_height,
+            .next_block_height(),
         2
     );
 
@@ -383,7 +384,204 @@ async fn should_not_set_shared_pre_state_to_lower_block_height() {
             .execution_pre_state
             .lock()
             .unwrap()
-            .next_block_height,
+            .next_block_height(),
         next_block_height
     );
+}
+
+#[cfg(test)]
+mod trie_chunking_tests {
+    use std::sync::Arc;
+
+    use casper_execution_engine::engine_state::engine_config::{
+        DEFAULT_FEE_HANDLING, DEFAULT_REFUND_HANDLING,
+    };
+    use casper_storage::global_state::{
+        state::StateProvider,
+        trie::{Pointer, Trie},
+    };
+    use casper_types::{
+        account::AccountHash,
+        bytesrepr,
+        execution::{Transform, TransformKind},
+        testing::TestRng,
+        ActivationPoint, CLValue, Chainspec, ChunkWithProof, CoreConfig, Digest, EraId, Key,
+        ProtocolConfig, StoredValue, TimeDiff,
+    };
+    use prometheus::Registry;
+    use tempfile::tempdir;
+
+    use crate::{
+        components::fetcher::FetchResponse,
+        contract_runtime::ContractRuntimeError,
+        types::{ChunkingError, TrieOrChunk, TrieOrChunkId, ValueOrChunk},
+    };
+
+    use super::{Config as ContractRuntimeConfig, ContractRuntime};
+
+    #[derive(Debug, Clone)]
+    struct TestPair(Key, StoredValue);
+
+    // Creates the test pairs that contain data of size
+    // greater than the chunk limit.
+    fn create_test_pairs_with_large_data() -> [TestPair; 2] {
+        let val = CLValue::from_t(
+            String::from_utf8(vec![b'a'; ChunkWithProof::CHUNK_SIZE_BYTES * 2]).unwrap(),
+        )
+        .unwrap();
+        [
+            TestPair(
+                Key::Account(AccountHash::new([1_u8; 32])),
+                StoredValue::CLValue(val.clone()),
+            ),
+            TestPair(
+                Key::Account(AccountHash::new([2_u8; 32])),
+                StoredValue::CLValue(val),
+            ),
+        ]
+    }
+
+    fn extract_next_hash_from_trie(trie_or_chunk: TrieOrChunk) -> Digest {
+        let next_hash = if let TrieOrChunk::Value(trie_bytes) = trie_or_chunk {
+            if let Trie::Node { pointer_block } = bytesrepr::deserialize::<Trie<Key, StoredValue>>(
+                trie_bytes.into_inner().into_inner().into(),
+            )
+            .expect("Could not parse trie bytes")
+            {
+                if pointer_block.child_count() == 0 {
+                    panic!("expected children");
+                }
+                let (_, ptr) = pointer_block.as_indexed_pointers().next().unwrap();
+                match ptr {
+                    Pointer::LeafPointer(ptr) | Pointer::NodePointer(ptr) => ptr,
+                }
+            } else {
+                panic!("expected `Node`");
+            }
+        } else {
+            panic!("expected `Trie`");
+        };
+        next_hash
+    }
+
+    // Creates a test ContractRuntime and feeds the underlying GlobalState with `test_pair`.
+    // Returns [`ContractRuntime`] instance and the new Merkle root after applying the `test_pair`.
+    fn create_test_state(rng: &mut TestRng, test_pair: [TestPair; 2]) -> (ContractRuntime, Digest) {
+        let temp_dir = tempdir().unwrap();
+        let chainspec = Chainspec {
+            protocol_config: ProtocolConfig {
+                activation_point: ActivationPoint::EraId(EraId::from(2)),
+                ..ProtocolConfig::random(rng)
+            },
+            core_config: CoreConfig {
+                max_associated_keys: 10,
+                max_runtime_call_stack_height: 10,
+                minimum_delegation_amount: 10,
+                prune_batch_size: 5,
+                strict_argument_checking: true,
+                vesting_schedule_period: TimeDiff::from_millis(1),
+                max_delegators_per_validator: 0,
+                allow_auction_bids: true,
+                allow_unrestricted_transfers: true,
+                fee_handling: DEFAULT_FEE_HANDLING,
+                refund_handling: DEFAULT_REFUND_HANDLING,
+                ..CoreConfig::random(rng)
+            },
+            wasm_config: Default::default(),
+            system_costs_config: Default::default(),
+            ..Chainspec::random(rng)
+        };
+        let contract_runtime = ContractRuntime::new(
+            temp_dir.path(),
+            &ContractRuntimeConfig::default(),
+            Arc::new(chainspec),
+            &Registry::default(),
+        )
+        .unwrap();
+        let empty_state_root = contract_runtime.engine_state().get_state().empty_root();
+        let mut effects = casper_types::execution::Effects::new();
+        for TestPair(key, value) in test_pair {
+            effects.push(Transform::new(key, TransformKind::Write(value)));
+        }
+        let post_state_hash = contract_runtime
+            .engine_state()
+            .commit_effects(empty_state_root, effects)
+            .expect("applying effects to succeed");
+        (contract_runtime, post_state_hash)
+    }
+
+    fn read_trie(contract_runtime: &ContractRuntime, id: TrieOrChunkId) -> TrieOrChunk {
+        let serialized_id = bincode::serialize(&id).unwrap();
+        match contract_runtime
+            .fetch_trie_local(&serialized_id)
+            .expect("expected a successful read")
+        {
+            FetchResponse::Fetched(found) => found,
+            FetchResponse::NotProvided(_) | FetchResponse::NotFound(_) => {
+                panic!("expected to find the trie")
+            }
+        }
+    }
+
+    #[test]
+    fn returns_trie_or_chunk() {
+        let rng = &mut TestRng::new();
+        let (contract_runtime, root_hash) =
+            create_test_state(rng, create_test_pairs_with_large_data());
+
+        // Expect `Trie` with NodePointer when asking with a root hash.
+        let trie = read_trie(&contract_runtime, TrieOrChunkId(0, root_hash));
+        assert!(matches!(trie, ValueOrChunk::Value(_)));
+
+        // Expect another `Trie` with two LeafPointers.
+        let trie = read_trie(
+            &contract_runtime,
+            TrieOrChunkId(0, extract_next_hash_from_trie(trie)),
+        );
+        assert!(matches!(trie, TrieOrChunk::Value(_)));
+
+        // Now, the next hash will point to the actual leaf, which as we expect
+        // contains large data, so we expect to get `ChunkWithProof`.
+        let hash = extract_next_hash_from_trie(trie);
+        let chunk = match read_trie(&contract_runtime, TrieOrChunkId(0, hash)) {
+            TrieOrChunk::ChunkWithProof(chunk) => chunk,
+            other => panic!("expected ChunkWithProof, got {:?}", other),
+        };
+
+        assert_eq!(chunk.proof().root_hash(), hash);
+
+        // try to read all the chunks
+        let count = chunk.proof().count();
+        let mut chunks = vec![chunk];
+        for i in 1..count {
+            let chunk = match read_trie(&contract_runtime, TrieOrChunkId(i, hash)) {
+                TrieOrChunk::ChunkWithProof(chunk) => chunk,
+                other => panic!("expected ChunkWithProof, got {:?}", other),
+            };
+            chunks.push(chunk);
+        }
+
+        // there should be no chunk with index `count`
+        let serialized_id = bincode::serialize(&TrieOrChunkId(count, hash)).unwrap();
+        assert!(matches!(
+            contract_runtime.fetch_trie_local(&serialized_id),
+            Err(ContractRuntimeError::ChunkingError(
+                ChunkingError::MerkleConstruction(_)
+            ))
+        ));
+
+        // all chunks should be valid
+        assert!(chunks.iter().all(|chunk| chunk.verify().is_ok()));
+
+        let data: Vec<u8> = chunks
+            .into_iter()
+            .flat_map(|chunk| chunk.into_chunk())
+            .collect();
+
+        let trie: Trie<Key, StoredValue> =
+            bytesrepr::deserialize(data).expect("trie should deserialize correctly");
+
+        // should be deserialized to a leaf
+        assert!(matches!(trie, Trie::Leaf { .. }));
+    }
 }
